refactor: 融合const/filters & mobile模板中的filters
