--- template/frameworks/constant/const/filter.js
@@ -1,8 +1,0 @@
-import dayjs from 'dayjs'
-
-export function formatDate(time, format = 'YYYY-MM-DD') {
-  if (!time && time !== 0) {
-    return ''
-  }
-  return dayjs(time).format(format)
-}
